refactor: reorganize middleware structure and update imports for consistency
