+++ server/src/controllers/classroom-membership.controller.ts
@@ -1,7 +1,7 @@
 import type { Response } from "express";
 import { ClassroomMembershipService } from "services/application/ClassroomMembershipService";
-import { errorResponse } from "utils";
 import type { EnhancedAuthRequest } from "../types/infrastructure/express/requests";
+import { errorResponse } from "../utils/errors";
 
 // Initialize service
 const membershipService = new ClassroomMembershipService();
