+++ server/src/middleware/auth/rateLimiter.ts
@@ -2,7 +2,7 @@ import rateLimit from "express-rate-limit";
 
 export const loginLimiter = rateLimit({
   windowMs: 60 * 1000,
-  handler: (req, res) => {
+  handler: (_, res) => {
     res.status(429).json({ message: "Too many requests. Please try again later." });
   },
 });
