+++ server/src/middleware/validation/validateRequest.ts
@@ -1,12 +1,14 @@
 import type { Request, Response, NextFunction } from "express";
-import { errorResponse } from "utils";
 import { type AnyZodObject, ZodError } from "zod";
+import { errorResponse } from "../../utils/errors";
 
+// Factory Pattern
 /**
  * Middleware that validates request data against a Zod schema
  * @param schema - The Zod schema to validate against
  * @param source - The request property to validate (body, query, params)
  */
+// This is a factory function that creates middleware
 export const validateRequest =
   (schema: AnyZodObject, source: "body" | "query" | "params" = "body") =>
   async (req: Request, res: Response, next: NextFunction) => {
