+++ server/src/routes/user.router.ts
@@ -1,8 +1,8 @@
 import express from "express";
 import { getCurrentUser, getUser, updateUser } from "controllers/user.controller";
-import { uploadMultiple } from "middlewares/upload/uploadMiddleware";
-import { withAuth } from "./middleware/classroom.middleware";
-import { withEnhancedAuth } from "middlewares/auth/enhancedAuth";
+import { uploadMultiple } from "middleware/upload/uploadMiddleware";
+import { withAuth } from "../middleware/domain/classroom.middleware";
+import { withEnhancedAuth } from "middleware/auth/enhancedAuth";
 
 const router = express.Router();
