--- server/src/utils/index.ts
@@ -1,15 +1,0 @@
-// JWT utilities
-export * from "./jwt";
-
-// Error classes
-export * from "./errors";
-
-// Response formatters
-export * from "./responses";
-
-// Validation utilities
-export * from "./validators";
-
-// Middleware utilities
-export * from "./middleware/authorization";
-export * from "./middleware/controller";
