+++ server/src/utils/responses.ts
@@ -1,19 +1,6 @@
 import type { Response } from "express";
 
-export const errorResponse = (
-  res: Response,
-  message = "An error occurred",
-  status = 500,
-  details?: any
-) => {
-  res.status(status).json({
-    success: false,
-    message,
-    ...(details && { details }),
-  });
-};
-
-// todo: add it
+// todo: Add it to the controllers and put here error handling also
 export const successResponse = (
   res: Response,
   data: any,
