+++ server/src/validations/index.ts
@@ -1,0 +1,5 @@
+export * from "./schemas/auth.schema";
+
+export * from "./schemas/classroom.schema";
+
+export { validateHandle } from "./validators/handle";
